% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsg_count.R
\name{vsg_count}
\alias{vsg_count}
\title{Count the number of VSGs in a group.}
\usage{
vsg_count(df, ..., keep_all = FALSE)
}
\arguments{
\item{df}{A data frame, usually containing VSGSeq results data.}

\item{...}{A set of columns to 'group_by' and count the number of VSGs. These
arguments are passed into dplyr::group_by function, then a dplyr::count
function is applied.}

\item{keep_all}{A logical argument. TRUE keeps all the rows and columns
originally put into the function.}
}
\value{
Returns a date frame with the VSG counts based on the input group_by
arguments.
}
\description{
Count the number of VSGs expressed in a designated group of parameters from a
VSGSeq output results table.
}
\examples{
vsg_count(results_df, mouse, tissue, day)
vsg_count(results_df, mouse, tissue, day, keep_all = TRUE)
}
