% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsg_distinct_vectorize.R
\name{vsg_distinct_vectorize}
\alias{vsg_distinct_vectorize}
\title{Return the distinct objects of a column as a vector.}
\usage{
vsg_distinct_vectorize(df, ...)
}
\arguments{
\item{df}{A data frame, usually containing VSGSeq results data.}

\item{...}{A set of columns to pass to dplyr::distinct().}
}
\value{
Returns a vector of the distinct objects in the column designated.
}
\description{
This function helps the vsg_expand function.
}
\examples{
vsg_distinct_vectorize(df, col1)
}
