% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsg_expand.R
\name{vsg_expand}
\alias{vsg_expand}
\title{Creates all possible combinations of sample ~ VSG (col1 ~ col2).}
\usage{
vsg_expand(df, col1, col2, merge = TRUE)
}
\arguments{
\item{df}{A data frame, usually containing VSGSeq results data.}

\item{col1}{The first column to expand on, this is typically the samplename
column. (Needs to be in quotes)}

\item{col2}{The second column to expand with, this is typically the VSG column
with the name of each VSG.}

\item{merge}{A logical argument to merge back with the original data to get percent info in the output.}
}
\value{
Returns a date frame with the VSG data for every possible combination of sample ~ VSG (col1 ~ col2)
}
\description{
This takes a VSG seq output data frame and expands it to include all possible
combinations of sample ~ VSG (col1 ~ col2). Then resets all Percent values to
be zero if a VSG was not found in a sample.
}
\examples{
vsg_expand(results_df, "samplename", "VSG")
}
