% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsg_max.R
\name{vsg_max}
\alias{vsg_max}
\title{Find the highest expressed VSG}
\usage{
vsg_max(df, ...)
}
\arguments{
\item{df}{A data frame, usually containing VSGSeq results data.}

\item{...}{A set of columns to 'group_by'. These arguments are passed into
dplyr::group_by function.}
}
\value{
Returns a date frame with the rows of VSGs with the maximum Percent
value based on the group_by arguments.
}
\description{
Find the most highly expressed VSG in a designated group.
}
\examples{
vsg_max(results_df, mouse, tissue, day)
}
