% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsg_split_mdt.R
\name{vsg_split_mdt}
\alias{vsg_split_mdt}
\title{Splits a samplename column into three columns, mouse, day, and tissue (mdt).}
\usage{
vsg_split_mdt(df, col_to_split, by = "_", convert = TRUE)
}
\arguments{
\item{df}{A data frame, usually containing VSGSeq results data.}

\item{col_to_split}{The column to split, usually a samplename column.}

\item{by}{The character to split by. The default is an underscore ("_").}

\item{convert}{If TRUE, will run type.convert() with as.is = TRUE on new
columns. This is useful if the component columns are integer, numeric or
logical.}
}
\value{
Returns a date frame with the designate column split into mouse, day, and tissue columns.
}
\description{
I kept having to do this so I made a function to split my usualy samplename
nomenclature into the designated mouse, day, adn tissue columns (mdt).
}
\examples{
vsg_split_mdt("samplename")
}
