% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsgseqtools.R
\docType{package}
\name{vsgseqtools}
\alias{vsgseqtools}
\title{vsgseqtools: Helpful Functions for VSG Sequencing Analysis}
\description{
A collection of functions that helps analyze and tidy up the output from VSG
sequencing.
}
\section{vsgseqtools functions}{

vsg_read, vsg_rename
}

